requirements.txt
openai
numpy
PyMuPDF
boto3
psycopg2-binary
ollama
python-dotenv 